When loading a layout, defer setting marks to the end.
If we set a mark as soon as we read it in the layout, it could be
that the "layout" key hasn't been read yet, which then causes i3
to crash due to an assertion error checking the layout not to be
L_DEFAULT when dumping the container. This happens when setting the
mark because we want to emit an event containing the container
structure.

We fix this by only memorizing the marks we saw in the layout and
only set them once the container has finished parsing.

fixes #2511